+++ haystack/document_stores/sql.py
@@ -34,7 +34,7 @@ except (ImportError, ModuleNotFoundError) as ie:
     _optional_component_not_installed(__name__, "sql", ie)
 
 
-from haystack.utils.import_utils import is_imported
+from haystack.utils.generalimport import is_imported
 from haystack.schema import Document, Label, Answer
 from haystack.document_stores.base import BaseDocumentStore, FilterType
 from haystack.document_stores.filter_utils import LogicalFilterClause
